[tanzania] move geometry handling functions to geometries.py
